backend/project/requirements.txt
Django==2.1.3
djangorestframework==3.9.0
virtualenv==16.0.0
